api_project/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
Django==3.2
django-mssql==1.8
djangorestframework==3.14.0
idna==3.4
mssql-django==1.3
pyodbc==4.0.39
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
